Se crea Clase para manejar los libros
