+++ src/entities/models/index.js
@@ -1,0 +1,4 @@
+const LibraryModel = require("./library");
+const BookModel = require("./book");
+
+module.exports = { LibraryModel, BookModel };
