+++ src/entities/models/library.js
@@ -1,32 +1,27 @@
 const { DataTypes } = require("sequelize");
 const { sequelize: db } = require("../../config/db-config");
-const Libreria = require("./libreria");
+const Libro = require("./book");
 
-const Libro = db.define("libro", {
+const Libreria = db.define("libreria", {
   id: {
     type: DataTypes.INTEGER,
     primaryKey: true,
     autoIncrement: true,
   },
-  isbn: {
-    type: DataTypes.INTEGER,
-    unique: true,
-    allowNull: false,
-  },
-  title: {
+  name: {
     type: DataTypes.STRING,
     allowNull: false,
   },
-  author: {
+  location: {
     type: DataTypes.STRING,
     allowNull: false,
   },
-  year: {
+  phone: {
     type: DataTypes.STRING,
     allowNull: false,
   },
 });
 
-Libro.belongsTo(Libreria, { foreignKey: "libraryId" });
+Libreria.hasMany(Libro, { foreignKey: "libraryId" });
 
-module.exports = Libro;
+module.exports = Libreria;
